Revert "Splits reporting into frontend and backend implementation"
