+++ .gitignore
@@ -2,5 +2,3 @@
 /_build/
 /.cproject
 /.project
-/.venv
-/.vscode
