+++ examples/CMakeLists.txt
@@ -10,4 +10,3 @@ endmacro()
 
 run_test(smoke)
 run_test(smoke_report)
-run_test(smoke_logger)
