+++ examples/smoke.cc
@@ -41,9 +41,9 @@ SC_MODULE (test) {
         uint64_t myint = mid + mid;
         myint += mid;
         if (mid == 0x450) {
-            SCP_INFO("ext test") << "Success";
+            SC_REPORT_INFO("ext test", "Success");
         } else {
-            SCP_INFO("ext test") << "Failure";
+            SC_REPORT_INFO("ext test", "Failour");
         }
     }
 };
