+++ examples/smoke_report.cc
@@ -23,6 +23,8 @@
 
 #include <cstdio>
 #include <string>
+#include <fstream>
+#include <streambuf>
 #include <unistd.h>
 
 SC_MODULE (test4) {
@@ -142,6 +144,14 @@ int sc_main(int argc, char** argv) {
 
     std::string logfile = "/tmp/scp_smoke_report_test." +
                           std::to_string(getpid());
+    scp::init_logging(
+        scp::LogConfig()
+            .logLevel(scp::log::DEBUG) // set log level to debug
+            .msgTypeFieldWidth(20)
+            .fileInfoFrom(5)
+            .logAsync(false)
+            .printSimTime(false)
+            .logFileName(logfile)); // make the msg type column a bit tighter
     SCP_INFO() << "Constructing design";
     test toptest("top");
     test1 t1("t1");
@@ -150,5 +160,56 @@ int sc_main(int argc, char** argv) {
     sc_core::sc_start();
     SCP_WARN() << "Ending simulation";
 
-    return 0;
+#ifdef FMT_SHARED
+    std::string fmtstr = "FMT String : Cached version default";
+#else
+    std::string fmtstr = "Please add FMT library for FMT support.";
+#endif
+
+    std::string expected =
+        R"([    info] [                0 s ]SystemC             : Constructing design
+[    info] [                0 s ]out.class           : constructor
+[ warning] [                0 s ]out.class           : constructor
+[   debug] [                0 s ]top                 : First part
+[    info] [                0 s ]top                 : top
+[    info] [                0 s ]top                 : top->top->top
+[   debug] [                0 s ]top                 : Second part
+[    info] [                0 s ]ext test            : Success
+[    info] [                0 s ]SystemC             : Uncached version empty
+[    info] [                0 s ]top                 : )" +
+        fmtstr + R"(
+[    info] [                0 s ]top                 : UnCached version feature using SCMOD macro
+[    info] [                0 s ]top                 : Cached version using (m_my_logger)
+[    info] [                0 s ]top                 : Cached version with D
+[    info] [                0 s ]t1.t2.t3_1          :  .  T3 D Logger "other" "feature.one"
+[ warning] [                0 s ]t1.t2.t3_1          :  .  T3 D Logger "other" "feature.one"
+[    info] [                0 s ]t1.t2.t3_1          :  .  T3 Logger ()
+[ warning] [                0 s ]t1.t2.t3_1          :  .  T3 Logger ()
+[    info] [                0 s ]t1.t2.t3_2          :  .  T3 D Logger "other" "feature.one"
+[ warning] [                0 s ]t1.t2.t3_2          :  .  T3 D Logger "other" "feature.one"
+[ warning] [                0 s ]t1.t2.t3_2          :  .  T3 Logger ()
+[    info] [                0 s ]t1.t2.t4            :  .   T4 Logger() 1
+[ warning] [                0 s ]t1.t2.t4            :  .   T4 Logger() 1
+[    info] [                0 s ]t1.t2.t4            :  .   T4 Logger() 2
+[ warning] [                0 s ]t1.t2.t4            :  .   T4 Logger() 2
+[ warning] [                0 s ]t1.t2               :   T2 Logger()
+[ warning] [                0 s ]My.Name             :  T1 My.Name typed log
+[ warning] [                0 s ]t1                  :  T1 Logger()
+[    info] [                0 s ]t1                  : Thing1?
+[ warning] [                0 s ]t1                  : Thing1?
+[ warning] [                0 s ]t1                  : Thing2?
+[    info] [                0 s ]SystemC             : Starting simulation
+[ warning] [                0 s ]SystemC             : Ending simulation
+)";
+
+    std::ifstream lf(logfile);
+    std::string out((std::istreambuf_iterator<char>(lf)),
+                    std::istreambuf_iterator<char>());
+
+    std::cout << "out file\n" << out << "\n";
+    std::cout << "expected\n" << expected << "\n";
+    std::cout << "Number of difference: " << out.compare(expected) << "\n";
+
+    std::remove(logfile.c_str());
+    return out.compare(expected);
 }
